new: delete read and delete progress functionalities

+++ src/use-cases/delete-follow-use-case.ts
@@ -1,14 +1,21 @@
 import { FollowsRepository } from "@/repositories/follows-repository";
 
+import { UnauthorizedError } from "./_errors/unauthorized-error";
+
 interface DeleteFollowUseCaseRequest {
     followerId: string;
     followingId: string;
+    userId: string;
 }
 
 export class DeleteFollowUseCase {
     constructor(private followsRepository: FollowsRepository) {}
 
-    async execute({ followerId, followingId }: DeleteFollowUseCaseRequest): Promise<void> {
+    async execute({ followerId, followingId, userId }: DeleteFollowUseCaseRequest): Promise<void> {
+        if (followerId !== userId) {
+            throw new UnauthorizedError();
+        }
+
         return await this.followsRepository.delete({
             followerId,
             followingId,
